Create app.js
This is a rudimentary Twitter Bot that is supposed to pull 10 tweets, any of which can fall between Jan 25th 2019 and the current date, that contain the name "Michael Cohen" within them, push the text of those tweets onto an array and then retweet said tweets randomly from that array. Currently I am running into a bug that prevents the bot from pulling tweets, so the bot is currently tweeting 'undefined', meaning it's not tweeting anything.